New inner class for GUI text display added. GUI is fully functional!
