+++ src/codeu/chat/util/GUI.java
@@ -70,6 +70,13 @@ public class GUI {
         rootButtonPanel.setLayout(new GridLayout(1, 6));
         rootInputPanel.add(rootButtonPanel);
         JButton rootHelp = new JButton("Help");
+        rootHelp.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent arg0) {
+                chat.handleCommand("help");
+            }
+
+        });
         rootButtonPanel.add(rootHelp);
         JButton listUsers = new JButton("List Users");
         listUsers.addActionListener(new ActionListener() {
@@ -130,7 +137,6 @@ public class GUI {
             public void actionPerformed(ActionEvent arg0) {
                 chat.handleCommand("u-sign-in " + rootTextInput.getText());
                 panelSwitcher.show(switchPanel, "userPanel");
-                rootOutput.close();
                 System.setOut(userOutput);
                 System.setErr(userOutput);
             }
@@ -248,7 +254,6 @@ public class GUI {
             @Override
             public void actionPerformed(ActionEvent arg0) {
                 chat.handleCommand("back");
-                userOutput.close();
                 System.setErr(rootOutput);
                 System.setOut(rootOutput);
                 panelSwitcher.show(switchPanel, "rootPanel");
@@ -281,7 +286,6 @@ public class GUI {
             @Override
             public void actionPerformed(ActionEvent e) {
                 chat.handleCommand("c-add " + userTextInput.getText());
-                userOutput.close();
                 System.setOut(convoOutput);
                 System.setErr(convoOutput);
                 panelSwitcher.show(switchPanel, "convoPanel");
@@ -293,7 +297,6 @@ public class GUI {
             @Override
             public void actionPerformed(ActionEvent arg0) {
                 chat.handleCommand("c-join " + userTextInput.getText());
-                userOutput.close();
                 System.setOut(convoOutput);
                 System.setErr(convoOutput);
                 panelSwitcher.show(switchPanel, "convoPanel");
@@ -386,7 +389,6 @@ public class GUI {
             @Override
             public void actionPerformed(ActionEvent arg0) {
                 chat.handleCommand("back");
-                convoOutput.close();
                 System.setOut(userOutput);
                 System.setErr(userOutput);
                 panelSwitcher.show(switchPanel, "userPanel");
@@ -401,5 +403,23 @@ public class GUI {
         window.pack();
         window.setVisible(true);
         window.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+    }
+}
+
+/**
+ * Created by Jiahui Chen on 8/2/2017.
+ * Creates a custom output stream that will display to the GUI's text panel.
+ */
+class GUIOutputDisplay extends OutputStream {
+    private JTextArea textDisplay;
+
+    public GUIOutputDisplay(JTextArea textArea){
+        textDisplay = textArea;
+    }
+
+    //redirects console/terminal output to JTextArea
+    @Override
+    public void write(int b) throws IOException {
+        textDisplay.append(String.valueOf((char) b));
     }
 }
